refactor(copyright): split refiner and extraction modules
Move large refiner regex tables and pattern extract helpers into focused sibling modules so the copyright code is easier to navigate without changing behavior.
