+++ src/copyright/refiner/authors_junk_patterns.rs
@@ -1,0 +1,109 @@
+// SPDX-FileCopyrightText: Provenant contributors
+// SPDX-License-Identifier: Apache-2.0
+
+use std::sync::LazyLock;
+
+use regex::Regex;
+
+pub(super) static AUTHORS_JUNK_PATTERNS: LazyLock<Vec<Regex>> = LazyLock::new(|| {
+    let patterns = [
+        r"(?i)\bpromote products derived from\b",
+        r"(?i)\bendorse or promote\b",
+        r"(?i)^the builder\b",
+        r"(?i)^the line highlight\b",
+        r"(?i)^the initial developer\b",
+        r"(?i)^trademark\b",
+        r"(?i)^time to time\b",
+        r"(?i)^the group of people\b",
+        r"(?i)^by,? or\b",
+        r"(?i)^group conversion\b",
+        r"(?i)^ctnewmethod\b",
+        r"(?i)\bplugins?\. fixes\b",
+        r"(?i)\bnormalized to upper\b",
+        r"(?i)\benhancing and supporting\b",
+        r"(?i)\band to credit the\b",
+        r"(?i)^other promise\b",
+        r"(?i)^record factory\b",
+        r"(?i)^the object\b",
+        r"(?i)^the owner,?\b",
+        r"(?i)^the job$",
+        r"(?i)^the ietf\b",
+        r"(?i)^manually\b",
+        r"(?i)^register\b",
+        r"(?i)^communication sent\b",
+        r"(?i)^developers tom\b",
+        r"(?i)^distribute\s+contributors$",
+        r"(?i)^ext4\.\b",
+        r"(?i)\bmore documentation\b",
+        r"(?i)\breturn enum\b",
+        r"(?i)\breturn u\d",
+        r"(?i)\bmore details of status\b",
+        r"(?i)\bu64$",
+        r"(?i)^\d+\.\d+\s+\d+-\w+-\d+\s+fix\b",
+        r"(?i)\bbut not limited to communication\b",
+        r"(?i)\bfor have helping\b",
+        r"(?i)\bunit of \d+mb\b",
+        r"(?i)\bfor the\s+[\p{L}0-9._+-]+(?:\s+[\p{L}0-9._+-]+){0,2}\s+project\b",
+        r"(?i)\bTenable licenses\b",
+        r"(?i)\bwith participation of the open\b",
+        r"(?i)\bfurthermore\b",
+        r"(?i)\bits cell\. we\b",
+        r"(?i)\b@version \$id\b",
+        r"(?i)\bsymbols viewer\b",
+        r"(?i)\bfinal specification itself\b",
+        r"(?i)\bfor each of the audio\b",
+        r"(?i)\bfrom start to end\b",
+        r"(?i)\boperator to\b",
+        r"(?i)^programming with objects\b",
+        r"(?i)^grateful to\b",
+        r"(?i)^would also like to thank\b",
+        r"(?i)^would like to thank\b",
+        r"(?i)^date modified$",
+        r"(?i)^date header id name\b",
+        r"(?i)^technical committee$",
+        r"(?i)^users of the program$",
+        r"(?i)^should not be interpreted\b",
+        r"(?i)^works devices national\b",
+        r"(?i)\band its \d+\.\s*neither\b",
+        r"(?i)\band its effective immediately\b",
+        r"(?i)\band its neither the\b",
+        r"(?i)^effective immediately\b",
+        r"(?i)^(?:\d+\.)?\s*neither\b",
+        r"(?i)\blastmod\b.*\bstream.*\bregisterfield\b",
+        r"(?i)^who\s+hopes\b",
+        r"(?i)^so preceded by\b",
+        r"(?i)^bounce, so we\b",
+        r"(?i)^the current user$",
+        r"(?i)^select\s+\.\w+$",
+        r"(?i)^with inkscape\b",
+        r"(?i)^in my book\b",
+        r"(?i)^chebyshev approximation to the$",
+        r"(?i)^developers,\s+testers,\s+and\s+maintainers\b",
+        r"(?i)^maintainers\.\s+when making changes to\b",
+        r"(?i)^point to the$",
+        r"(?i)^did not confirm that$",
+        r"(?i)^could not find a direct reference to$",
+        r"(?i)^developers aware of the problem and plan to fix it$",
+        r"(?i)^with the help of\s+.+$",
+        r"(?i)^with the mode of \d{3,4}$",
+        r"(?i)^kernel afs\.\s+skip afs metadata and acls$",
+        r"(?i)^with a (?:fsf|dco)$",
+        r"(?i)^gives unlimited$",
+        r"(?i)^word assigns past and future changes\b",
+        r"(?i)^maintainers\s*<[^>]+>\s+from\s+https?://\S+$",
+        r"(?i)^versions,\s+and$",
+        r"(?i)^versions$",
+        r"(?i)^makes$",
+        r"(?i)^grants irrevocable$",
+        r"(?i)^version information$",
+        r"(?i)^not responsible for the consequences of use of$",
+        r"(?i)^at sunpro, a sun microsystems, inc\. business$",
+        r"(?i)^debian by .+\bupstream-contact\b.*$",
+        r"(?i)^was originally written by .+\bit is now maintained by\b.*$",
+        r"(?i)^writing to the appropriate\b.*$",
+        r"(?i)^the task'?s numa mempolicy\b.*$",
+        r"(?i)^the authors laboriously took the trouble\b.*$",
+        r"(?i)^laboriously took the trouble\b.*$",
+    ];
+    patterns.iter().filter_map(|p| Regex::new(p).ok()).collect()
+});
