+++ src/copyright/refiner/holders_junk_patterns.rs
@@ -1,0 +1,699 @@
+// SPDX-FileCopyrightText: Provenant contributors
+// SPDX-License-Identifier: Apache-2.0
+
+use std::sync::LazyLock;
+
+use regex::Regex;
+
+/// Regex patterns for junk holder detections (license boilerplate fragments).
+pub(super) static HOLDERS_JUNK_PATTERNS: LazyLock<Vec<Regex>> = LazyLock::new(|| {
+    let patterns = [
+        r"(?i)^licenses?,\s+and/or\b",
+        r"(?i)^holders?,\s*authors\b",
+        r"(?i)^notice,\s+and\b",
+        r"(?i)^notice,\s+in\b",
+        r"(?i)^but\s+its\s+distribution\b",
+        r"(?i)\bprovided\s+to\s+the\s+public\s+domain\b",
+        r"(?i)^patent\s+or\s+trademark\b",
+        r"(?i)^notice,\s+but\s+the\s+BSD,\s+MIT\s+and\s+UoI/NCSA\s+licenses\s+do\s+not\b",
+        r"(?i)\bVALUE\s+OriginalFilename\b",
+        r"(?i)\bOriginalFilename\b",
+        r"(?i)\bEND\s*$",
+        r"(?i)^etc\.?\s+in\b",
+        r"(?i)^version\s*,\s*etc\b",
+        r"(?i)^\(d\),\s*\d+(?:\.\d+)*\.?$",
+        r"(?i)\bliable for\b",
+        r"(?i)\bappear in all copies\b",
+        r"(?i)\bdisclaimer of warranty\b",
+        r"(?i)\bdisclaimer for the program\b",
+        r"(?i)\bit may be distributed\b",
+        r"(?i)\bwho places the program\b",
+        r"(?i)\bkeep intact all the\b",
+        r"(?i)\bshall not be used in advertising\b",
+        r"(?i)\bpromote the sale\b",
+        r"(?i)\bpromote products derived\b",
+        r"(?i)\bother dealings in\b",
+        r"(?i)\bhas been advised of the possibility\b",
+        r"(?i)\bfailure of essential purpose\b",
+        r"(?i)\bthe licenses? granted in\b",
+        r"(?i)\bcovering the original code\b",
+        r"(?i)\bwithout notice from apple\b",
+        r"(?i)\bcompletely and accurately document\b",
+        r"(?i)\bother proprietary\b",
+        r"(?i)\bpatent rights?\b",
+        r"(?i)^patents?\s*,\s*trade secrets?\b",
+        r"(?i)^patent\s*,\s*or trademark\b",
+        r"(?i)^patent\s*,\s*trademark\b",
+        r"(?i)^trade secrets?\b",
+        r"(?i)^including\s+without\s+limitation\b",
+        r"(?i)^trademarks?$",
+        r"(?i)\bincluding.{0,10}but not limited\b",
+        r"(?i)\bincluding your\b",
+        r"(?i)\bincluding the\b",
+        r"(?i)\bcopyrighted material\b",
+        r"(?i)\bselected patent\b",
+        r"(?i)\bin the work\b",
+        r"(?i)\bin the document\b",
+        r"(?i)\bthe original work\b",
+        r"(?i)\bpermit and encourage\b",
+        r"(?i)\bpermitted copying\b",
+        r"(?i)\bto do the following\b",
+        r"(?i)\bas a result of\b",
+        r"(?i)\breinstated permanently\b",
+        r"(?i)\breinstated\b",
+        r"(?i)\bexplicitly and finally terminates\b",
+        r"(?i)\bfails to notify\b",
+        r"(?i)\bnotifies\b",
+        r"(?i)\bthe above\b",
+        r"(?i)\bthe software,?$",
+        r"(?i)^software$",
+        r"(?i)\bsuspend your rights\b",
+        r"(?i)\bderivative works\b",
+        r"(?i)\bpublicly display\b",
+        r"(?i)\bpublicly perform\b",
+        r"(?i)\bof competent jurisdiction\b",
+        r"(?i)\bexceptions and limitations\b",
+        r"(?i)\bfair use\b",
+        r"(?i)\bfair dealing\b",
+        r"(?i)\btreaty adopted\b",
+        r"(?i)\breflecting the\b",
+        r"(?i)^\d{1,2}:\d{2}\s+[a-z0-9][a-z0-9._-]*$",
+        r"(?i)\bappears? in\b",
+        r"(?i)\bsaying\b.*\bdistributed\b",
+        r"(?i)\bif the item a binary\b",
+        r"(?i)\bone digital image or graphic\b",
+        r"(?i)\bperceptible, measurable\b",
+        r"(?i)\bthe entire\b",
+        r"(?i)\bsemblance of artistic control\b",
+        r"(?i)\bcommercially reasonable efforts\b",
+        r"(?i)\bto endorse or promote\b",
+        r"(?i)\bimmediately at the beginning\b",
+        r"(?i)\bunmodified\b",
+        r"(?i)\beasier identification\b",
+        r"(?i)\b(l?gpl|lgpl) group\b",
+        r"(?i)^symbol in\b",
+        r"(?i)^trademark$",
+        r"(?i)^printf\b",
+        r"(?i)^the top level of\b",
+        r"(?i)^the following\b",
+        r"(?i)^the resulting\b",
+        r"(?i)^whoever named in\b",
+        r"(?i)^as specified below\b",
+        r"(?i)^\((?:r|c|tm)\)\s*,\s*(?:&\s*\d{2,4}\s*,\s*){5,}.*$",
+        r"(?i)^not used to limit\b",
+        r"(?i)^the coordinator$",
+        r"(?i)^provided\b",
+        r"(?i)^provides the work\b",
+        r"(?i)^final\s+[a-z_][a-zA-Z0-9_]*$",
+        r"(?i)\bthis\.[a-zA-Z]",
+        r"(?i):function\b",
+        r"^(?:[a-z_][a-zA-Z0-9_]*\.[A-Za-z_][A-Za-z0-9_]*\([^)]*\)|[a-z_][a-zA-Z0-9_]*\.[A-Za-z_]*[A-Z][A-Za-z0-9_]*)$",
+        r"(?i)\bm\. y\. name\b",
+        r"(?i)^version of nameif\b",
+        r"(?i)\bunless explicitly identified\b",
+        r"(?i)^version 3 of the$",
+        // Holder false positives from license boilerplate
+        r"(?i)\b(if any) with\b",
+        r"(?i)^(d),\b",
+        r"(?i)\bas a market\b",
+        r"(?i)\bprocedures\b",
+        r"(?i)\bcollectively\b",
+        r"(?i)\bgiving your\b",
+        r"(?i)\bspecified addresses\b",
+        r"(?i)^the base\b",
+        r"(?i)^the library\b",
+        r"(?i)\bthe library,\b",
+        r"(?i)\bthe library among\b",
+        r"(?i)\breferences to\b",
+        r"(?i)\bstating\b.*\bdistributed\b",
+        r"(?i)^terminate\b",
+        r"(?i)\beffective immediately\b",
+        r"(?i)^keep intact\b",
+        r"(?i)^material outside\b",
+        r"(?i)\bsaying\b",
+        // Trailing legal text patterns
+        r"(?i)\bdistributed under\b",
+        r"(?i)\blicensed under\b",
+        r"(?i)\bthe terms\b.*\blicense\b",
+        r"(?i)\bthe standard version\b",
+        // Code-like patterns in holders
+        r"(?i)\bif\s*\(",
+        r"(?i)\bfunction\s*\(",
+        r"(?i)\breturn\b.*\bfunction\b",
+        r"(?i)\bvar\s+\w",
+        r"(?i)\bthis\.\w+\(",
+        // Trailing text patterns in holders
+        r"(?i)\bCredited\b",
+        r"(?i)\bConverted to\b",
+        r"(?i)\breworked by\b",
+        r"(?i)\bVarious bits\b",
+        r"(?i)\bCopying and distribution\b",
+        r"(?i)\bGPLd\b",
+        r"(?i)\bLicense-Alias\b",
+        r"(?i)\bWritten by\b",
+        r"(?i)\bModified by the\b",
+        r"(?i)\btitle Legal\b",
+        r"(?i)\bContact <",
+        r"(?i)\b- Placed\b",
+        r"(?i)\bUnder the terms\b",
+        r"(?i)\binfo have been\b",
+        r"(?i)\bicon support\b",
+        r"(?i)\bfull list\b",
+        r"(?i)^proprietary$",
+        r"(?i)^not limited to\b",
+        r"(?i)\bprocurement of substitute goods or services\b",
+        r"(?i)^related rights\b",
+        r"(?i)^related or neighboring\b",
+        r"(?i)^was owned solely by fsf$",
+        r"(?i)^years may be listed$",
+        r"(?i)^notice\s*,\s*license\s+and\s+disclaimer\.?$",
+        r"(?i)^trademarks?\s*,\s*trade\s+secrets?\b",
+        r"(?i)^the\s+standard$",
+        r"(?i)^the\s+product$",
+        r"(?i)^rsa$",
+        r"(?i)^xerox corporation$",
+        r"(?i)^david j\. bradshaw$",
+        r"(?i)^gias kay lee$",
+        r"(?i)^tim ruffles$",
+        r"(?i)^products\.\s+tenable licenses$",
+        r"(?i)^the Embedded Configurable Operating System\.?$",
+        r"(?i)^(?:convert|multiply)\s+(?:a\s+)?(?:chebyshev|hermite|laguerre|legendre)\b",
+        r"(?i)^treaties$",
+        r"(?i)\bpatent\s+or\s+other\s+licenses\s+necessary\b",
+        r"(?i)\bMODULEAUTHOR\b",
+        r"(?i)^THE\s+[\p{L}0-9._-]+'S\b",
+        r"(?i)^the\s+(?:[\p{L}0-9._-]+\s+){1,5}operating\s+system\.?$",
+        r"(?i)\bpkg\.(author|homepage)\b",
+        r"(?i)\bdate\.year\b",
+        r"(?i)\bYYYY-MM-DD\b",
+        r"(?i)<\s*pkg\.[a-zA-Z0-9_.-]+\s*>",
+        r"(?i)^http://www\.quirksmode\b",
+        r"[→⟶]",
+        r"(?i)\bSTATEMENTS AND\b",
+        r"(?i)\bAS IS$",
+        r"(?i)\bAS IS CONDITION\b",
+        r"(?i)\bNOTICES OR THIS\b",
+        r"(?i)\bDOCUMENTATION ISC\b",
+        r"(?i)\bpixmaps svg\b",
+        r"(?i)\bFull text of\b",
+        r"(?i)\btransferred to Nokia\b",
+        r"(?i)\bAS PER APPLICABLE\b",
+        r"(?i)\bSection 105\b",
+        r"(?i)\bGNU AGPL\b",
+        r"(?i)\bagreement with the\b",
+        r"(?i)\bgives Customer\b",
+        r"(?i)\bshall mean\b",
+        r"(?i)\bEnterprise Edition\b",
+        r"(?i)\bContributing Authors\b",
+        r"(?i)\bAll Downstream\b",
+        r"(?i)\bSource Code to\b",
+        r"(?i)\bPROTECTION AND IS\b",
+        r"(?i)\bnot removed\b",
+        r"(?i)\bversion 3\.1 of\b",
+        r"(?i)\bGPL version\b",
+        r"(?i)\bCopyright/g\b",
+        r"(?i)\bdata/c\.m4\b",
+        r"(?i)\binside so it\b",
+        r"(?i)\bmatch standard format\b",
+        r"(?i)\bin each output\b",
+        r"(?i)\bstr::npos\b",
+        r"(?i)\btimes in xrange\b",
+        r"(?i)\bin zlib\.h\b",
+        r"(?i)\ball paragraphs\b",
+        r"(?i)\buse, copy, modify\b",
+        r"(?i)\bdistribute it with\b",
+        r"(?i)\bother intellectual property\b",
+        r"(?i)\btreaties\. Title\b",
+        r"(?i)\bexempting the\b",
+        r"(?i)\bwith exception of\b",
+        r"(?i)\bas noted in the\b",
+        r"(?i)\bThe Product is\b",
+        r"(?i)\bThe arguments as\b",
+        r"(?i)\bpertaining to distribution\b",
+        r"(?i)\bVERBATIM\b",
+        r"(?i)\bintact$",
+        r"(?i)\binformation\.\b",
+        r"(?i)\bdoing$",
+        r"(?i)^holders,? but\b",
+        r"(?i)^its author\b",
+        r"(?i)^in its\b",
+        r"(?i)^in the\b",
+        r"(?i)^offer\b",
+        r"(?i)^copy the\b",
+        r"(?i)^owned by\b",
+        r"(?i)^the team$",
+        r"(?i)^the project$",
+        r"(?i)^the republic of\b",
+        r"(?i)^the google\b",
+        r"(?i)^the jetty\b",
+        r"(?i)^the acknowledgment\b",
+        r"(?i)^the combination of\b",
+        r"(?i)^the lavantech\b",
+        r"(?i)^all source code\b",
+        r"(?i)^all translated\b",
+        r"(?i)^all the rich\b",
+        r"(?i)^author,? or contributor\b",
+        r"(?i)^authors,? and contributors\b",
+        r"(?i)^its authors\b",
+        r"(?i)^its cell\b",
+        r"(?i)^automatically without\b",
+        r"(?i)^more information\b",
+        r"(?i)^infringement can\b",
+        r"(?i)^header of\b",
+        r"(?i)^const (group|projects)\b",
+        r"(?i)^there clear\b",
+        r"(?i)^things like\b",
+        r"(?i)^custom nessus\b",
+        r"(?i)^whenever reasonably\b",
+        r"(?i)^gnu general\b",
+        r"(?i)^general public\b",
+        r"(?i)^man page\b",
+        r"(?i)^merged arm\b",
+        r"(?i)^tcl/tk policy\b",
+        r"(?i)^in license\b",
+        r"(?i)^law,? \b",
+        r"(?i)^license,? to the\b",
+        r"(?i)^s status\b",
+        r"(?i)^as i developed\b",
+        r"(?i)^improved syntax\b",
+        r"(?i)^inclusion in\b",
+        r"(?i)^disclaim all\b",
+        r"(?i)^directly copied\b",
+        r"(?i)^as found in\b",
+        r"(?i)^years updated\b",
+        r"(?i)\bcontrol over the development\b",
+        r"(?i)\bartistic control\b",
+        r"(?i)\bcompilation not used to limit\b",
+        r"(?i)\blegal rights of the compilation\b",
+        r"(?i)\bindividual works permit\b",
+        r"(?i)\bDocument included in\b",
+        r"(?i)\blocated in .* and .* located in\b",
+        r"(?i)\binternational treaty\b",
+        r"(?i)\bapplicable$",
+        r"(?i)\bcontrat et tous\b",
+        r"(?i)\ben anglais\b",
+        r"(?i)\bdocuments connexes\b",
+        r"(?i)^seek a different\b",
+        r"(?i)^sign so\b",
+        r"(?i)^like sta\b",
+        r"(?i)^page i/o\b",
+        r"(?i)^\(mon tue\b",
+        r"(?i)^gt\. zero\b",
+        r"(?i)^with recursive\b",
+        r"(?i)^ecfieldelement\b",
+        r"(?i)^setresultsname\b",
+        r"(?i)^semanticdirection\b",
+        r"(?i)^content ssense\b",
+        r"(?i)^attr value\b",
+        r"(?i)^match\(ident\)\b",
+        r"(?i)^assert\.equal\b",
+        r"(?i)^h\.matches\b",
+        r"(?i)^bd\(b\.\b",
+        r"(?i)^b\(an\)\d",
+        r"(?i)^b\(ase\b",
+        r"(?i)^b\(onstant\b",
+        r"(?i)^g\(al\)\b",
+        r"(?i)^y fj\b",
+        r"(?i)^y fp\b",
+        r"(?i)^u r\(\d",
+        r"(?i)^u q\d",
+        r"(?i)^y rd\b",
+        r"(?i)^y aey\b",
+        r"(?i)^as z$",
+        r"(?i)^i\. uao\b",
+        r"(?i)^e qugu\b",
+        r"(?i)^bj d\b",
+        r"(?i)^cj d\b",
+        r"(?i)^dj d\b",
+        r"(?i)^jj d\b",
+        r"(?i)^objc,? bp\b",
+        r"(?i)^10 a - 10$",
+        r"(?i)^b a, b$",
+        r"(?i)^unknown [a-z]{1,3}$",
+        r"(?i)^unknown [a-z]\d\b",
+        r"^[a-z]{1,2} [a-z]{1,2}$",
+        r"^[A-Z][a-z] [A-Z]$",
+        r"^[a-z][A-Z] [A-Z]{1,2}$",
+        r"(?i)^ato\b.*\bae\b",
+        r"(?i)^xz\b.*\bijr\b",
+        r"(?i)^zigd\d\b",
+        r"(?i)^yy thq\b",
+        r"(?i)^ss'ss",
+        r"(?i)^pspsy\b",
+        r"(?i)^oo2\b",
+        r"(?i)^c/ps\b",
+        r"(?i)^cn:class\b",
+        r"(?i)^c2001\b",
+        r"(?i)^ocoo\b",
+        r"(?i)^a!\b",
+        r"(?i)^aae\b",
+        r"(?i)^a\(r\)\b",
+        r"(?i)^deg,?\b.*deg\b",
+        r"(?i)^cii1/4\b",
+        r"(?i)^vo u\d",
+        r"(?i)^ul\b",
+        r"(?i)^xl\b",
+        r"(?i)^wl\b",
+        r"(?i)^crarr\b",
+        r"(?i)^x\$\?\b",
+        r"(?i)^e\$\?\b",
+        r"(?i)^length\?null\b",
+        r"(?i)^c\.warn\b",
+        r"(?i)^b\.status\b",
+        r"(?i)^as\(c,\b",
+        r"(?i)^cc\.fr$",
+        r"(?i)^q ltd$",
+        r"(?i)^zbar\b",
+        r"(?i)^ssssy$",
+        r"(?i)^ssss$",
+        r"(?i)^as5$",
+        r"(?i)^r'b$",
+        r"(?i)^\?12$",
+        r"(?i)^tj et\b",
+        r"(?i)^adobe.*\bairtm\b",
+        r"(?i)^adobe.*\bair\u{2122}\b",
+        r"[\x00-\x1f]",
+        r"°°°",
+        r"ÿÿÿ",
+        r"\u{9a}f",
+        r"\u{96}b",
+        r"\u{9d}v",
+        r"^[A-Z][a-z]$",
+        r"^[A-Z][b-z]$",
+        r"^[a-z][A-Z]$",
+        r"^holder\.\b",
+        r"^holder,\b",
+        r"^holders,\b",
+        r"^holder as\b",
+        r"(?i)^applied to\b",
+        r"(?i)^designation\b",
+        r"(?i)^registered$",
+        r"(?i)^component$",
+        r"(?i)^count$",
+        r"(?i)^group$",
+        r"(?i)^isupper$",
+        r"(?i)^folded$",
+        r"(?i)^dean$",
+        r"(?i)^targetpath$",
+        r"(?i)^libre-software$",
+        r"(?i)^\(2\)\.\s*if\b",
+        r"(?i)^\(as found in\b",
+        r"(?i)^\(directly copied\b",
+        r"(?i)^\(if any\)\b",
+        r"(?i)^m\(h",
+        r"(?i)^b\(onsisting\b",
+        r"(?i)^inria-enpc\b",
+        r"(?i)^uossoo\b",
+        r"(?i)^ocothd\b",
+        r"(?i)^otho\b",
+        r"(?i)^iouoi\b",
+        r"(?i)^aiuaey\b",
+        r"(?i)^aoth\b",
+        r"(?i)^ato\b",
+        r"(?i)^aaeamooa\b",
+        r"(?i)^eeiaeiaaoa\b",
+        r"(?i)^exauauuao\b",
+        r"(?i)^erxa\d",
+        r"(?i)^ijax\b",
+        r"(?i)^u1e\b",
+        r"(?i)^degu\b",
+        r"(?i)^xmlns\b",
+        r"(?i)^\u{201c}adobe\b",
+        r"(?i)\bthe resulting\b",
+        r"(?i)\ball source code included in\b",
+        r"(?i)\bsource code distributed need not\b",
+        r"(?i)\bdo not make\b",
+        r"(?i)\bgive all recipients\b",
+        r"(?i)\brecipients of the\b",
+        r"(?i)\bmay be used\b",
+        r"(?i)\bthe accompanying\b",
+        r"(?i)\bas represented by no\b",
+        r"(?i)^compute hessian\b",
+        r"(?i)^nat\d+\.is",
+        r"(?i)^a\.compatibility\b",
+        r"(?i)^retained at the\b",
+        r"(?i)^timer code\b",
+        r"(?i)^an sr-iov\b",
+        r"(?i)^applies to the regex\b",
+        r"(?i)^apple's sf pro\b",
+        r"(?i)^xmlns:\?",
+        r"(?i)^swfobject\b",
+        r"(?i)^program\b.*\btalke studio\b",
+        r"(?i)^footer$",
+        r"(?i)^legal\s+notice(?:\s+div\s+div\.legalnotice\s+p)?$",
+        r"(?i)^color\s+666666$",
+        r"(?i)^debian\b.*\bjames troup\b",
+        r"(?i)^\$id\$\b",
+        r"(?i)^the uc berkeley\b",
+        r"(?i)^ococoa\b",
+        r"(?i)^grant\. i\b",
+        r"(?i)^the gnome libraries\b",
+        r"(?i)^as is group\b",
+        r"(?i)^match\(ident\)\s*ast\b",
+        r"(?i)^holder,? author,? or contributor\b",
+        r"(?i)^holders,? authors,? and contributors\b",
+        r"(?i)\bportions of\b",
+        r"(?i)\bsome parts of\b",
+        r"(?i)\bthe source$",
+        r"(?i)\bthe source code\b",
+        r"(?i)\b\. the source\b",
+        r"(?i)^p b i n do$",
+        r"(?i)^tue \w+ \d+ \w+ \w+ -",
+        r"(?i)^info for$",
+        r"(?i)^material,? only\b",
+        r"(?i)^(d),? \d",
+        r"(?i)^c\. schmidt$",
+        r"(?i)^gdb free software\b",
+        r"(?i)^va$",
+        r"(?i)^wing$",
+        r"(?i)^hillion$",
+        r"(?i)^(TOUPPER|isascii|iscntrl|isprint|yyunput|ambiguous|TRUE FALSE)$",
+        r"(?i)^(width|len|do|date|year|note|update|notive|all the)$",
+        r"(?i)^undef\s+\w+$",
+        r"(?i)^i\.e\.,\b",
+        r"(?i)^endif\b",
+        r"(?i)^definedummyword\b",
+        r"(?i)^register int\b",
+        r"(?i)^l \(unsigned\b",
+        r"(?i)^\(\(unsigned\b",
+        r"(?i)^notices all\b",
+        r"(?i)^may not be removed\b",
+        r"(?i)^duplicated in\b",
+        r"(?i)^copyright for a\b",
+        r"(?i)^copyright info\b",
+        r"(?i)^COPYRIGHT HOLDERS AS\b",
+        r"(?i)^Mouse Wheel Support\b",
+        r"(?i)^Joseph Gil avalable\b",
+        r"(?i)^Original code for Bayer\b",
+        r"(?i)@remark Read the",
+        r"(?i)\bEND END$",
+        r"(?i)^inc\.,\s*Id Software\b",
+        r"(?i)^Id Software.*Id Software\b",
+        r"(?i)\bavalab?le at\b",
+        r"(?i)^\(k \d+ k \d+\b",
+        r"(?i)^\(unsigned char\)\b",
+        r"(?i)^\(int\) TOUPPER\b",
+        r"(?i)^(isascii|isdigit|isalpha|isalnum|isupper|islower|isspace|isgraph|ispunct|isxdigit)\b",
+        r"(?i)^ungetc\b",
+        r"(?i)^yylval\b",
+        r"(?i)^symrec\b",
+        r"(?i)^arg\s*\+\+",
+        r"(?i)^gunichar\b",
+        r"(?i)^TRUE FALSE$",
+        r"(?i)^undef\b",
+        r"(?i)^0 1$",
+        r"(?i)^\d{1,2}$",
+        r"(?i)^ok-for-header$",
+        r"(?i)^date\b.*\bDon't assume\b",
+        r"(?i)^notive in the\b",
+        r"(?i)\bdon't assume ascii\b",
+        r"(?i)^all the$",
+        r"(?i)\bftp://\b",
+        r"(?i)^CC Computer Consultants\b.*\bContact\b",
+        r"(?i)^16 \(\(d\)\b",
+        r"(?i)^\(c\) s-$",
+        r"(?i)^z \?$",
+        r"(?i)^Z \?$",
+        r"(?i)^this-\s*setStencil\b",
+        r"(?i)^temp\d+$",
+        r"(?i)^table\.set\b",
+        r"(?i)^strict!?\s*-?\d",
+        r"(?i)^slen$",
+        r"(?i)^save to iv\b",
+        r"(?i)^r,?\s*div\b",
+        r"(?i)^r sround\b",
+        r"(?i)^r \(s\)$",
+        r"(?i)^put chain\b",
+        r"(?i)^problem,?\s*work-around\b",
+        r"(?i)^prec prec\b",
+        r"(?i)^pr this\b",
+        r"(?i)^Paul Rusty Russell\b.*\bPlaced\b",
+        r"(?i)^Paul Mackerras\b.*\bpipe read\b",
+        r"(?i)^packet$",
+        r"(?i)^p can be called\b",
+        // ICS false positives: code fragments, boilerplate, gibberish
+        r"(?i)^the parens part of\b",
+        r"(?i)^i\.e\.\s*,?\s*call the\b",
+        r"(?i)^8 \(\(b\)\b",
+        r"(?i)^\d+ \(trail\b",
+        r"(?i)^\d+ illegal\b",
+        r"(?i)^strict!\s",
+        r"(?i)^\(s\) \(i\)$",
+        r"(?i)^0x[0-9a-fA-F]+",
+        r"(?i)^\d+ \+0x",
+        r"(?i)^\d+ &0x",
+        r"(?i)^it a lead surrogate\b",
+        r"(?i)^uint\d+$",
+        r"(?i)^Construct a set of\b",
+        r"(?i)^clause removed\b",
+        r"(?i)^0x\d",
+        r"(?i)^below\.?\s*(Please)?\b",
+        r"(?i)^below$",
+        r"(?i)^above$",
+        r"(?i)^\(qbuf\b",
+        r"(?i)^applies to code\b",
+        r"(?i)^0x7f\b",
+        r"(?i)^\(with the right granted\b",
+        r"(?i)^fi$",
+        r"(?i)^as follows$",
+        r"(?i)^dst$",
+        r"(?i)^dst-\s",
+        r"(?i)^< 0 e->",
+        r"(?i)^\(\d+ \(pattern\b",
+        r"(?i)^ACM and IEEE\b",
+        r"(?i)^make it clear$",
+        r"(?i)^ifdef$",
+        r"(?i)^exploring the\b.*\bcultural\b",
+        r"(?i)^do prec\b",
+        r"(?i)^EOF &&\b",
+        r"(?i)^4\+\(r\)$",
+        r"(?i)^c&0x",
+        r"(?i)^cp\)$",
+        r"(?i)^ctype$",
+        r"(?i)^macroptr\b",
+        r"(?i)^\(shf\)\b",
+        r"(?i)^MAGIC$",
+        r"(?i)^out\.ro$",
+        r"(?i)^attribution$",
+        r"(?i)^\d+ ,\s*l\b.*\b(unsigned|endif)\b",
+        r"(?i)^asm bswapl\b",
+        r"(?i)^d'\(l\)$",
+        r"(?i)^l endif$",
+        r"(?i)^03o$",
+        r"(?i)^apply$",
+        r"(?i)^\(cp\)$",
+        r"(?i)^buflen\b.*\bbuf$",
+        r"(?i)^buffer\s+[a-z]$",
+        r"(?i)^etc\b.*\bstrings\b",
+        r"(?i)^1,\s*cls\.\b",
+        r"(?i)^this-\s*set\w+\b",
+        r"(?i)^\(scale\s+\d\)\s+\d",
+        r"(?i)^fWidth$",
+        r"(?i)^dst \d$",
+        r"(?i)^i \(s\)\b.*\b(while|endif)\b",
+        r"(?i)^i ,?\s*div\b",
+        r"(?i)^i sround\b",
+        r"(?i)^i,?\s*s while\b",
+        r"(?i)^r,?\s*div\b",
+        r"(?i)^r,?\s*s$",
+        r"(?i)^\(a\) \(b\)$",
+        r"(?i)^FILE\.*\s+\w+\.\w+\s+AUTHOR\b",
+        r"(?i)^Accumulate$",
+        r"(?i)^ctable\b",
+        r"(?i)^CREDITS PORTING\b",
+        r"(?i)^24 endif$",
+        r"(?i)^\^ 0x\d",
+        r"(?i)^putchar\b.*\bputchar\b",
+        r"(?i)^do while$",
+        r"(?i)^\^ \(b\)$",
+        r"(?i)^help$",
+        r"(?i)^in gzlog\.h\b",
+        r"(?i)^decoded by\b",
+        r"(?i)^IBM Corporation\.$",
+        r"(?i)^Lotus Development Corporation\.$",
+        r"(?i)^disclaimer$",
+        r"(?i)^disclaims all\b",
+        r"(?i)^Foundation IBM\b",
+        r"(?i)^http://sizzlejs\b",
+        r"(?i)^Kasım$",
+        r"(?i)^Akim Demaille$",
+        r"(?i)^Joel E\. Denny$",
+        r"(?i)^num \d$",
+        r"(?i)^letters\b.*\bc - A\b",
+        r"(?i)^classify$",
+        r"(?i)^\(r\) l l$",
+        r"(?i)^ok letters\b.*\bcond\b",
+        r"(?i)^flags ptbl-\b",
+        r"(?i)^holders,?\s*disclaims\b",
+        r"(?i)^\d+L$",
+        r"(?i)^\d+L,\s*l\b",
+        r"(?i)^Chain has\b",
+        r"(?i)^DEBUGP\b",
+        r"(?i)^Only user$",
+        r"(?i)^cindex chains$",
+        r"(?i)^foot-\s*target\b",
+        r"(?i)^OProfile authors\b.*@remark",
+        r"(?i)^and are$",
+        r"(?i)^mea-\s*setOffset\b",
+        r"(?i)^meta-\s*registerClass\b",
+        r"(?i)^EOF &&\s",
+        r"(?i)^Bit8u$",
+        r"(?i)^cvPoint3D32f$",
+        r"(?i)^G2P ADJ\b",
+        r"(?i)^info to be inserted\b",
+        r"(?i)^0 isupper$",
+        r"(?i)^0-9,- \d",
+        r"(?i)^97 static$",
+        r"(?i)^8 \(\(b\)\s*\d",
+        r"(?i)^\d+ \(\(d\)\s*\d",
+        r"(?i)^\d+ \(\(a\)\s*\d",
+        r"(?i)\b@version \$Id\b",
+        r"(?i)^ds Status works\b",
+        r"(?i)^holder\.\s*AS\b",
+        r"(?i)^oCOOA\b",
+        r"(?i)^as\(c,\s*field\b",
+        r"(?i)^a!\s*\+-",
+        r"(?i)^\(xmlns:\?\s*\^",
+        r"(?i)^Tue\s+\w+\s+\d+\s+\w+\s+\w+\s+<",
+        r"(?i)\b\d{2}-[A-Z]{3}-\d{2}\s+Bugfixes\b",
+        r"(?i)\bpartial mlock\b",
+        r"(?i)\bskb\.\s*The buffer\b",
+        r"(?i)^IBM Corp\.\s*Auxtrace\b",
+        r"(?i)^digÃ",
+        r"(?i)^que le prÃ",
+        // Garbled binary data holders (from junk-copyright tests)
+        r"^AaACEEeUB",
+        r"^AaeaMOOAA\d",
+        r"^EEIaeIaAAOAE",
+        r"^OCOthDTh",
+        r"^YY ThQ",
+        r"^YY$",
+        r"(?i)^NIST\.\d+\.\d+\.",
+        r"(?i)AEEEUAU",
+        r"(?i)\$\?I\$\?i\$\?I",
+        r"^x!C/!O$",
+        r"(?i)^33,BD\(b\b",
+        r"(?i)^2AICAA",
+        r"(?i)^Rear\b",
+        r"(?i)^Rear Left$",
+        r"(?i)^Rear Right$",
+        r"(?i)^kavol$",
+        r"(?i)^assigned to the United States Government\b",
+        r"(?i)^so preceded by\b",
+        r"(?i)^bounce, so we\b",
+        r"(?i)^conditions,?\s+but\s+instead\b",
+        r"(?i)^www\.\S+",
+        r"(?i)^jornada\s+\d+$",
+        r"(?i)^mjander\b",
+        r"(?i)^notice,\s*a\s+notice\b",
+        r"(?i),\s*a\s+notice\b",
+        r"(?i)^\(scale\s+\d+\)$",
+        r"(?i)^scale\s+\d+$",
+        r"(?i)^i,\s*div\s+while$",
+        r"(?i)^isinstanceof$",
+        r"(?i)^contributor,\s*path$",
+        r"(?x)\b[A-Za-z_][A-Za-z0-9_]*\([^)]*\).*[A-Za-z_][A-Za-z0-9_]*\.[A-Za-z_][A-Za-z0-9_]*\(",
+        r"^(?:[\u{0080}-\u{00FF}]+\s*){6,}$",
+    ];
+    patterns.iter().filter_map(|p| Regex::new(p).ok()).collect()
+});
